slub: add support for verifying slab sanitization
This could support !PAGE_SANITIZE by zeroing in the object setup code in
those cases, but for now it takes the simple approach of depending on
page sanitization.

This is an extension to the sanitization feature in PaX for when
sacricifing more performance for security is acceptable.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>
(cherry picked from commit fc661f6b8fe3f9d980b9b1465e8f335c40951d0a)